Move "Find duplicate name" methods to a separate static class
